tweaks to policy automation form fields
