+++ frontend/pages/policies/edit/components/PolicyAutomations/PolicyAutomationsList.tests.tsx
@@ -1,0 +1,185 @@
+import React from "react";
+import { render, screen } from "@testing-library/react";
+
+import { IPolicy } from "interfaces/policy";
+import PolicyAutomationsList from "./PolicyAutomationsList";
+
+// Stub SoftwareIcon to avoid asset resolution in tests
+jest.mock("pages/SoftwarePage/components/icons/SoftwareIcon", () => {
+  return () => <span data-testid="software-icon" />;
+});
+
+const createMockPolicy = (overrides?: Partial<IPolicy>): IPolicy => ({
+  id: 1,
+  name: "Test policy",
+  query: "SELECT 1;",
+  description: "",
+  author_id: 1,
+  author_name: "Admin",
+  author_email: "admin@example.com",
+  resolution: "",
+  platform: "darwin",
+  team_id: 1,
+  created_at: "2026-01-01T00:00:00Z",
+  updated_at: "2026-01-01T00:00:00Z",
+  critical: false,
+  calendar_events_enabled: false,
+  conditional_access_enabled: false,
+  type: "dynamic",
+  ...overrides,
+});
+
+describe("PolicyAutomationsList", () => {
+  describe("automations list", () => {
+    it("shows empty state when no automations are configured", () => {
+      render(
+        <PolicyAutomationsList
+          storedPolicy={createMockPolicy()}
+          currentAutomatedPolicies={[]}
+        />
+      );
+
+      expect(screen.getByText("No automations")).toBeInTheDocument();
+    });
+
+    it("shows software automation row", () => {
+      render(
+        <PolicyAutomationsList
+          storedPolicy={createMockPolicy({
+            install_software: { name: "Zoom", software_title_id: 42 },
+          })}
+          currentAutomatedPolicies={[]}
+        />
+      );
+
+      expect(screen.getByText("Zoom")).toBeInTheDocument();
+      expect(screen.queryByText("No automations")).not.toBeInTheDocument();
+    });
+
+    it("shows script automation row", () => {
+      render(
+        <PolicyAutomationsList
+          storedPolicy={createMockPolicy({
+            run_script: { id: 1, name: "fix.sh" },
+          })}
+          currentAutomatedPolicies={[]}
+        />
+      );
+
+      expect(screen.getByText("fix.sh")).toBeInTheDocument();
+    });
+
+    it("shows calendar automation row", () => {
+      render(
+        <PolicyAutomationsList
+          storedPolicy={createMockPolicy({ calendar_events_enabled: true })}
+          currentAutomatedPolicies={[]}
+        />
+      );
+
+      expect(screen.getByText("Maintenance window")).toBeInTheDocument();
+    });
+
+    it("shows conditional access automation row", () => {
+      render(
+        <PolicyAutomationsList
+          storedPolicy={createMockPolicy({
+            conditional_access_enabled: true,
+          })}
+          currentAutomatedPolicies={[]}
+        />
+      );
+
+      expect(screen.getByText("Block single sign-on")).toBeInTheDocument();
+    });
+
+    it("shows 'Webhook' for other automation when otherAutomationType is webhook", () => {
+      render(
+        <PolicyAutomationsList
+          storedPolicy={createMockPolicy({ id: 1 })}
+          currentAutomatedPolicies={[1]}
+          otherAutomationType="webhook"
+        />
+      );
+
+      expect(screen.getByText("Webhook")).toBeInTheDocument();
+      expect(screen.queryByText("Ticket")).not.toBeInTheDocument();
+    });
+
+    it("shows 'Ticket' for other automation when otherAutomationType is ticket", () => {
+      render(
+        <PolicyAutomationsList
+          storedPolicy={createMockPolicy({ id: 1 })}
+          currentAutomatedPolicies={[1]}
+          otherAutomationType="ticket"
+        />
+      );
+
+      expect(screen.getByText("Ticket")).toBeInTheDocument();
+      expect(screen.queryByText("Webhook")).not.toBeInTheDocument();
+    });
+
+    it("shows 'Webhook or ticket' for other automation when otherAutomationType is not set", () => {
+      render(
+        <PolicyAutomationsList
+          storedPolicy={createMockPolicy({ id: 1 })}
+          currentAutomatedPolicies={[1]}
+        />
+      );
+
+      expect(screen.getByText("Webhook or ticket")).toBeInTheDocument();
+    });
+  });
+
+  describe("footer text", () => {
+    it("shows default footer text when continuous_automations_enabled is not set", () => {
+      render(
+        <PolicyAutomationsList
+          storedPolicy={createMockPolicy()}
+          currentAutomatedPolicies={[]}
+        />
+      );
+
+      expect(
+        screen.getByText(
+          "Automations run on a host's first failure, or when a host's response changes from pass to fail."
+        )
+      ).toBeInTheDocument();
+    });
+
+    it("shows continuous footer text when continuous_automations_enabled is true", () => {
+      render(
+        <PolicyAutomationsList
+          storedPolicy={createMockPolicy({
+            continuous_automations_enabled: true,
+          })}
+          currentAutomatedPolicies={[]}
+        />
+      );
+
+      expect(
+        screen.getByText(/Software and script automations run/)
+      ).toBeInTheDocument();
+      expect(screen.getByText("every time")).toBeInTheDocument();
+      expect(
+        screen.getByText(/All other automations run on a host's first failure/)
+      ).toBeInTheDocument();
+    });
+
+    it("shows footer text even in the empty state", () => {
+      render(
+        <PolicyAutomationsList
+          storedPolicy={createMockPolicy()}
+          currentAutomatedPolicies={[]}
+        />
+      );
+
+      expect(screen.getByText("No automations")).toBeInTheDocument();
+      expect(
+        screen.getByText(
+          "Automations run on a host's first failure, or when a host's response changes from pass to fail."
+        )
+      ).toBeInTheDocument();
+    });
+  });
+});
