+++ frontend/pages/policies/edit/components/PolicyAutomations/_styles.scss
@@ -1,34 +1,35 @@
-.policy-automations {
+.patch-automation-cta {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  box-sizing: border-box;
+  width: 100%;
+  padding: $pad-medium;
+  background-color: $ui-fleet-black-5;
+  border: 1px solid $ui-fleet-black-10;
+  border-radius: 8px;
+
+  &__label {
+    font-size: $x-small;
+    color: $ui-fleet-black-75;
+  }
+}
+
+.policy-automations-list {
   display: flex;
   flex-direction: column;
   gap: 0.5rem;
   font-size: $x-small;
 
-  &__cta-card {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: $pad-medium;
-    background-color: $ui-fleet-black-5;
-    border: 1px solid $ui-fleet-black-10;
-    border-radius: 8px;
-    margin-bottom: $pad-medium;
-  }
-
-  &__cta-label {
-    font-size: $x-small;
-    color: $ui-fleet-black-75;
+  &__header {
+    color: $core-fleet-black;
+    font-weight: $bold;
   }
 
   &__list {
     max-width: 600px;
     border: 1px solid $ui-fleet-black-10;
     border-radius: 8px;
-  }
-
-  &__list-label {
-    color: $core-fleet-black;
-    font-weight: $bold;
   }
 
   &__row {
@@ -52,7 +53,6 @@
     display: flex;
     align-items: center;
     gap: $pad-small;
-    font-size: $x-small;
     color: $ui-fleet-black-75;
 
     a {
