+++ frontend/pages/policies/edit/screens/QueryEditor.tsx
@@ -12,11 +12,7 @@ import debounce from "utilities/debounce";
 import deepDifference from "utilities/deep_difference";
 import { getPathWithQueryParams } from "utilities/url";
 import { getErrorReason } from "interfaces/errors";
-import {
-  IPolicyFormData,
-  IPolicy,
-  OtherAutomationType,
-} from "interfaces/policy";
+import { IPolicyFormData, IPolicy } from "interfaces/policy";
 
 import BackButton from "components/BackButton";
 import PolicyForm from "pages/policies/edit/components/PolicyForm";
@@ -37,8 +33,6 @@ interface IQueryEditorProps {
   onOpenSchemaSidebar: () => void;
   renderLiveQueryWarning: () => JSX.Element | null;
   teamIdForApi?: number;
-  currentAutomatedPolicies?: number[];
-  otherAutomationType?: OtherAutomationType;
 }
 
 const QueryEditor = ({
@@ -56,8 +50,6 @@ const QueryEditor = ({
   onOpenSchemaSidebar,
   renderLiveQueryWarning,
   teamIdForApi,
-  currentAutomatedPolicies,
-  otherAutomationType,
 }: IQueryEditorProps): JSX.Element | null => {
   const { currentUser, isPremiumTier, filteredPoliciesPath } = useContext(
     AppContext
@@ -301,8 +293,6 @@ const QueryEditor = ({
         onClickAutofillDescription={onClickAutofillDescription}
         onClickAutofillResolution={onClickAutofillResolution}
         resetAiAutofillData={() => setPolicyAutofillData(null)}
-        currentAutomatedPolicies={currentAutomatedPolicies || []}
-        otherAutomationType={otherAutomationType}
       />
     </div>
   );
